Bullet Position and Center Instantiation 2
Improved some part of the system that made the bullet instantiation at the shooting point based  on its center so that now will make the position better but also made now the rotation of the sprite based on the  on its origin which I put it at the center of the  using the angle between the shooting point and the mouse position. Next, only need is to adjust the position of the sprite center to be the same with the collider center. Also I used a circle collider for the bullet which is better than a rectangle