UI Button Class 3
Added a new Button Class type as the child of the UI Button class named UI Clicked Button that is a specific type of UI button. Also important, I made the first step of the Scene Manager System so that now I can transition between scenes in the game manually for now. It needs some improvements but the 2 scenes are working correctly with this Scene Manager System for now.